Fix FE Bugs completed

+++ legalia-webapp/src/lib/components/researchresult/ResultSentenza.svelte
@@ -3,11 +3,12 @@
   import { dateFormat } from "../misc/UsefulFunctions.svelte";
 
   export let sentenza:Sentenza;
-  //let sentenzaContentModal;
+  let sentenzaContentModal;
 </script>
 <!--role="button" tabindex="0" on:dblclick={sentenzaContentModal.showModal()}-->
 <div class="w-full flex items-center rounded-lg border-l-4 transition-colors duration-200 my-2 p-0">
-    <div class="flex-grow min-w-0 rounded-md py-3 px-2 bg-white text-black hover:bg-gray-200" style="min-width: 75%;">
+    <!-- svelte-ignore a11y-no-static-element-interactions -->
+    <div on:dblclick={sentenzaContentModal.showModal()} class="flex-grow min-w-0 rounded-md py-3 px-2 bg-white text-black hover:bg-gray-200" style="min-width: 75%;">
       <p class="text-md font-medium truncate text" style="width: 90%">
         {sentenza.description.split(" \n")[0]}
       </p>
@@ -22,17 +23,26 @@
     </div> 
 </div>
 
-<!--<dialog bind:this={sentenzaContentModal} class="modal">
-  <div class="rounded-lg bg-white text-black" style="width:75%;max-height: 75vh;">
-    <div class="px-10 pt-10">
-      <h1 class="text-lg font-bold mb-4">{sentenza.description}</h1>
-      <div class="overflow-y-scroll pr-4 text-justify" style="max-height: 300px">
-        {@html sentenza.content.replace(/\n/g, "<br>")}
-      </div>
+<dialog bind:this={sentenzaContentModal} class="modal">
+  <div class="modal-box max-w-2xl bg-white text-black">
+    <div class="flex items-center justify-between mb-4 border-b pb-2">
+      <h3 class="text-lg font-bold">
+        {sentenza.description.split(" \n")[0]} n. {sentenza.number}
+      </h3>
+      <form method="dialog">
+        <button class="btn btn-sm btn-circle btn-ghost">✕</button>
+      </form>
     </div>
-    <div><p class="text-sm text-center my-5">Premi ESC per uscire</p></div>
+    <h4 class="text-lg font-semibold text-primary mb-3">{sentenza.description.split(" \n")[1]}</h4>
+    <div class="overflow-y-auto pr-4 text-justify" style="max-height: 300px">
+      <!-- Main content -->
+      <div class="prose max-w-full mb-4">
+        <div class="bg-base-200 p-4 rounded-lg bg-neutral-200 text-black">
+          <p>{sentenza.content}</p>
+        </div>
+    </div>
   </div>
-</dialog>-->
+</dialog>
 
 <style>
   /* Stile della scrollbar (per Chrome, Edge e Safari) */
